Move get_user_info() to l.lib.user.
