+++ landscape/manager/customgraph.py
@@ -6,11 +6,11 @@ from twisted.internet.defer import fail, DeferredList, succeed
 from twisted.python.compat import iteritems
 
 from landscape.lib.scriptcontent import generate_script_hash
+from landscape.lib.user import get_user_info, UnknownUserError
 from landscape.accumulate import Accumulator
 from landscape.manager.plugin import ManagerPlugin
 from landscape.manager.scriptexecution import (
-    ProcessFailedError, ScriptRunnerMixin, ProcessTimeLimitReachedError,
-    get_user_info, UnknownUserError)
+    ProcessFailedError, ScriptRunnerMixin, ProcessTimeLimitReachedError)
 
 
 class StoreProxy(object):
