+++ landscape/manager/scriptexecution.py
@@ -4,7 +4,7 @@ Functionality for running arbitrary shell scripts.
 @var ALL_USERS: A token indicating all users should be allowed.
 """
 import os
-import pwd
+import os.path
 import tempfile
 import shutil
 
@@ -12,14 +12,14 @@ from twisted.internet.protocol import ProcessProtocol
 from twisted.internet.defer import (
     Deferred, fail, inlineCallbacks, returnValue, succeed)
 from twisted.internet.error import ProcessDone
-from twisted.python.compat import unicode, _PY3
+from twisted.python.compat import unicode
 
 from landscape import VERSION
 from landscape.constants import UBUNTU_PATH
-from landscape.lib.encoding import encode_if_needed
 from landscape.lib.fetch import fetch_async, HTTPCodeError
 from landscape.lib.persist import Persist
 from landscape.lib.scriptcontent import build_script
+from landscape.lib.user import get_user_info
 from landscape.manager.plugin import ManagerPlugin, SUCCEEDED, FAILED
 
 
@@ -27,41 +27,6 @@ ALL_USERS = object()
 TIMEOUT_RESULT = 102
 PROCESS_FAILED_RESULT = 103
 FETCH_ATTACHMENTS_FAILED_RESULT = 104
-
-
-class UnknownUserError(Exception):
-    pass
-
-
-def get_user_info(username=None):
-    uid = None
-    gid = None
-    path = None
-    if username is not None:
-        if _PY3:
-            username_str = username
-        else:
-            username_str = encode_if_needed(username)
-        try:
-            # XXX: We have a situation with the system default FS encoding with
-            # Python 3 here: We have to pass a string to pwd.getpwnam(), but if
-            # the default does not support unicode characters, a
-            # UnicodeEncodeError will be thrown. This edge case can be harmful,
-            # if the user was added with a less restrictive encoding active,
-            # and is now retrieved with LC_ALL=C for example, as it is during
-            # automatic test runs. This should not be a problem under normal
-            # circumstances. Alternatively, a different way of parsing
-            # /etc/passwd would have to be implemented. A simple
-            # locale.setlocale() to use UTF-8 was not successful.
-            info = pwd.getpwnam(username_str)
-        except (KeyError, UnicodeEncodeError):
-            raise UnknownUserError(u"Unknown user '%s'" % username)
-        uid = info.pw_uid
-        gid = info.pw_gid
-        path = info.pw_dir
-        if not os.path.exists(path):
-            path = "/"
-    return (uid, gid, path)
 
 
 class ProcessTimeLimitReachedError(Exception):
