+++ landscape/manager/tests/test_scriptexecution.py
@@ -14,10 +14,10 @@ from landscape import VERSION
 from landscape.lib.fetch import HTTPCodeError
 from landscape.lib.persist import Persist
 from landscape.lib.testing import StubProcessFactory, DummyProcess
+from landscape.lib.user import get_user_info, UnknownUserError
 from landscape.manager.scriptexecution import (
     ScriptExecutionPlugin, ProcessTimeLimitReachedError, PROCESS_FAILED_RESULT,
-    UBUNTU_PATH, get_user_info, UnknownInterpreterError, UnknownUserError,
-    FETCH_ATTACHMENTS_FAILED_RESULT)
+    UBUNTU_PATH, UnknownInterpreterError, FETCH_ATTACHMENTS_FAILED_RESULT)
 from landscape.manager.manager import SUCCEEDED, FAILED
 from landscape.tests.helpers import LandscapeTest, ManagerHelper
